+++ CHANGELOG.md
@@ -4,6 +4,15 @@
 
 ## [Unreleased]
 
+## [2.0.4] - 2023-08-10
+
+### Continue Renovations
+
+- Fixed EDT Error
+- Fixed Disposable Error
+- Added clear Workflow cache option
+- Fixed false positives at Syntax Highlighting
+
 ## [2.0.3] - 2023-08-10
 
 ### Continue Renovations